+++ plugins/onestep-rabbitmq/src/onestep_rabbitmq/__init__.py
@@ -4,9 +4,7 @@ from importlib.metadata import PackageNotFoundError, version as _package_version
 
 from .connector import RabbitMQConnector, RabbitMQDelivery, RabbitMQQueue
 from .resources import register_resources
-from .resilience import classify_rabbitmq_error, register_error_classifiers
-
-register_error_classifiers()
+from .resilience import classify_rabbitmq_error
 
 try:
     __version__ = _package_version("onestep-rabbitmq")
@@ -22,6 +20,5 @@ __all__ = [
     "__version__",
     "classify_rabbitmq_error",
     "register",
-    "register_error_classifiers",
     "register_resources",
 ]