+++ plugins/onestep-sqs/src/onestep_sqs/__init__.py
@@ -4,9 +4,7 @@ from importlib.metadata import PackageNotFoundError, version as _package_version
 
 from .connector import SQSConnector, SQSDelivery, SQSQueue
 from .resources import register_resources
-from .resilience import classify_sqs_error, register_error_classifiers
-
-register_error_classifiers()
+from .resilience import classify_sqs_error
 
 try:
     __version__ = _package_version("onestep-sqs")
@@ -22,6 +20,5 @@ __all__ = [
     "__version__",
     "classify_sqs_error",
     "register",
-    "register_error_classifiers",
     "register_resources",
 ]